Move scraping functions to scraper/fetcher.py
